% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_import_format.R
\name{import_format}
\alias{import_format}
\title{Import and format USGS water flow data}
\usage{
import_format(
  site = "15908000",
  start_date = "1996-01-01",
  end_date = "2020-12-31",
  skiplines = 34
)
}
\arguments{
\item{site}{a numeric unique USGS identifier for a water monitoring site}

\item{start_date}{the desired start date}

\item{end_date}{the desired end date}
}
\value{
a list containing a data frame of raw daily flow data and a data frame of summarized monthly flow data
}
\description{
Import and format USGS water flow data
}
\examples{
dontrun{
import_format()
}
}
