% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_analysis-kw_dunns.R
\name{kw_dunns}
\alias{kw_dunns}
\title{Kruskal-Wallis and Dunn's tests for monthly differences in water flow across time periods}
\usage{
kw_dunns(dat)
}
\arguments{
\item{dat}{a data frame of daily water flow data returned from \code{import_format()}}
}
\value{
a list containing Kruskal Wallis chi square test results and a data
frame containing Dunn's test results
}
\description{
Kruskal-Wallis and Dunn's tests for monthly differences in water flow across time periods
}
\examples{
dontun{
kw_dunns(dat)
}
}
