% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_report.R
\name{run_report}
\alias{run_report}
\title{Run a water flow report for the Sag River}
\usage{
run_report(outdir = "./reports")
}
\arguments{
\item{outdir}{the output directory to store the report}
}
\value{
a pdf report
}
\description{
Run a water flow report for the Sag River
}
\examples{
dontrun{
run_report()
}
}
